add back static_cast and add comment

+++ multi_physics/QED/include/picsar_qed/containers/picsar_tables.hpp
@@ -983,7 +983,13 @@ namespace details{
                 throw std::runtime_error("raw_data contains invalid data.");
             if(m_how_many_y <= 0)
                 throw std::runtime_error("raw_data contains invalid data.");
-            m_values = VectorType(m_how_many_x*m_how_many_y);
+
+            //static_cast<int> here has the sole purpose of avoiding to trigger
+            //a CodeQL check in CI (Multiplication result converted to larger type)
+            //Realistically, m_how_many_x and m_how_many_y will never be large
+            //enough to cause an overflow
+            m_values = VectorType(static_cast<int>(m_how_many_x*m_how_many_y));
+
             auto vals = serialization::get_n_out<RealType>(
                     it_raw_data,
                     m_how_many_x*m_how_many_y);
